+++ tests/cli_common/mod.rs
@@ -5,6 +5,12 @@ use std::io::Write;
 use std::path::PathBuf;
 use tempfile::{tempdir, tempdir_in, TempDir};
 
+#[cfg(windows)]
+use dunce::canonicalize;
+
+#[cfg(not(windows))]
+use std::fs::canonicalize;
+
 /// Use to test commands in temporary directories
 pub struct TestDir {
     current_dir: TempDir,
@@ -80,7 +86,8 @@ impl TestDir {
         S: AsRef<OsStr>,
     {
         let mut cmd = Self::init_cmd();
-        cmd.current_dir(&self.current_dir)
+        let canonical_current_dir = canonicalize(&self.current_dir).expect("canonical current dir");
+        cmd.current_dir(&canonical_current_dir)
             .args(args)
             .assert()
             .success();
@@ -98,7 +105,8 @@ impl TestDir {
         S: AsRef<OsStr>,
     {
         let mut cmd = Self::init_cmd();
-        cmd.current_dir(&self.current_dir)
+        let canonical_current_dir = canonicalize(&self.current_dir).expect("canonical current dir");
+        cmd.current_dir(&canonical_current_dir)
             .args(args)
             .assert()
             .failure()